% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{aoc_delete_conclusion}
\alias{aoc_delete_conclusion}
\title{Delete a conclusion post}
\usage{
aoc_delete_conclusion(year = NULL)
}
\arguments{
\item{year}{An integer representing the year, from 2015 to the current year. Defaults to the current year.}
}
\description{
Delete the introduction post with relative path \code{./YYYY/day/conclusion} (where \code{YYYY}
is the value of \code{year}), if it exists.
}
\examples{
\dontrun{aoc_delete_conclusion(2022)}
}
\seealso{
\code{\link[=aoc_new_conclusion]{aoc_new_conclusion()}}
}
