% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{aoc_delete_intro}
\alias{aoc_delete_intro}
\title{Delete an introduction post}
\usage{
aoc_delete_intro(year = NULL)
}
\arguments{
\item{year}{An integer representing the year, from 2015 to the current year. Defaults to the current year.}
}
\description{
Delete the introduction post with relative path \code{./YYYY/day/YYYY-intro} (where \code{YYYY}
is the value of \code{year}), if it exists.
}
\examples{
\dontrun{aoc_delete_introduction(2022)}
}
\seealso{
\code{\link[=aoc_new_intro]{aoc_new_intro()}} \code{\link[=aoc_new_year]{aoc_new_year()}} \code{\link[=aoc_delete_year]{aoc_delete_year()}}
}
