% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{aoc_delete_introduction}
\alias{aoc_delete_introduction}
\title{Delete an introduction post}
\usage{
aoc_delete_introduction(year = NULL)
}
\arguments{
\item{year}{An integer representing the year. Defaults to the current year.}
}
\description{
Delete an introduction post
}
\examples{
\dontrun{aoc_delete_introduction(2022)}
}
