% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{aoc_get_input}
\alias{aoc_get_input}
\title{Get and save the puzzle input}
\usage{
aoc_get_input(day, year = NULL)
}
\arguments{
\item{day}{An integer between 1 and 25}

\item{year}{An integer representing the year, from 2015 to the current year. Defaults to the current year.}
}
\value{
Returns, invisibly, a character string with the absolute path to the
input file.
}
\description{
Get the puzzle input from the Advent of Code website for a given day and year
and save it to a file. The file will be saved in the current working
directory with a relative path \code{"./YYYY/day/DD/input"}, where \code{YYYY} is the
value of the year parameter and \code{DD} is value of the day parameter. This path
echoes the URL structure of the Advent of Code website.
}
\details{
This function assumes that you have an account on the \href{https://adventofcode.com}{Advent of Code website}. For this function to work, you must set
the \code{ADVENT_SESSION} environment variable in your \code{.Renviron} file. For
guidance on how to find your session token, see
\url{https://github.com/dgrtwo/adventdrob/tree/main#installation}. Once you have
your session token, you can set the environment variable with
\code{\link[usethis:edit]{usethis::edit_r_environ()}}. This function is adapted from
\url{https://github.com/dgrtwo/adventdrob/blob/main/R/input.R}.
}
\examples{
\dontrun{aoc_get_input(1, 2022)}
}
