% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{aoc_input_data_frame}
\alias{aoc_input_data_frame}
\title{Read puzzle input as a data frame}
\usage{
aoc_input_data_frame(
  day,
  year = NULL,
  file = "input",
  class = c("tbl_df", "data.frame"),
  col_names = FALSE,
  show_col_types = FALSE,
  view = FALSE
)
}
\arguments{
\item{day}{An integer between 1 and 25}

\item{year}{An integer representing the year, from 2015 to the current year. Defaults to the current year.}

\item{file}{Character string. The name of a file that exists in the directory
\code{"./YYYY/day/DD"}, where \code{YYYY} and \code{DD} are the values of \code{year} and
\code{day}. Defaults to \code{"input"}.}

\item{class}{character. One of 'tbl_df' or 'data.frame'. Default is 'tbl_df'.}

\item{col_names}{logical. Only relevant if class is \code{tbl_df}. If \code{TRUE}, the
first row of the input file will be used as the column names. Default is
\code{FALSE}}

\item{show_col_types}{logical. If \code{TRUE}, the column types will be shown.
Default is \code{FALSE}}

\item{view}{logical. If \code{TRUE}, calls \code{\link[utils:View]{utils::View()}} to view the returned input.
Default is \code{FALSE}}
}
\value{
A data frame containing the puzzle input for the day and year.
}
\description{
Read in the puzzle input, or other file, as a data frame, one row per line.
It assumes that the file is stored in the directory
\code{"./YYYY/day/DD"}, where \code{YYYY} and \code{DD} are the values of \code{year} and \code{day}.
By default, the file name is \code{"input"}.
This file will exist in this location if the post was created using \code{\link[=aoc_new_day]{aoc_new_day()}}.
class is \code{tbl_df}, the reading is done by \code{\link[readr:read_table]{readr::read_table()}}. If the class
is \code{data.frame}, the reading is done by \code{\link[utils:read.table]{utils::read.table()}}.
The defaults have been chosen to suit typical Advent of Code puzzle input.
}
\examples{
\dontrun{aoc_input_data_frame(8, 2020)}
}
\seealso{
\code{\link[=aoc_input_vector]{aoc_input_vector()}} \code{\link[=aoc_input_matrix]{aoc_input_matrix()}}
}
