% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{aoc_input_matrix}
\alias{aoc_input_matrix}
\title{Read puzzle input as a matrix}
\usage{
aoc_input_matrix(
  day,
  year = NULL,
  file = "input",
  mode = c("character", "numeric"),
  split = "",
  view = FALSE
)
}
\arguments{
\item{day}{An integer between 1 and 25}

\item{year}{An integer representing the year, from 2015 to the current year. Defaults to the current year.}

\item{file}{Character string. The name of a file that exists in the directory
\code{"./YYYY/day/DD"}, where \code{YYYY} and \code{DD} are the values of \code{year} and
\code{day}. Defaults to \code{"input"}.}

\item{mode}{Character string. One of 'character' or 'numeric'. Default is
'character'.}

\item{split}{character. The string to split the input on. Default is \code{""},
i.e. one character per column}

\item{view}{logical. If \code{TRUE}, calls \code{\link[utils:View]{utils::View()}} to view the returned input.
Default is \code{FALSE}}
}
\value{
A matrix containing the puzzle input for the day and year.
}
\description{
Read in the puzzle input, or other file, as a matrix, one row per line, by
default one character per column. It assumes that the file is stored in the
directory \code{"./YYYY/day/DD"}, where \code{YYYY} and \code{DD} are the values of \code{year}
and \code{day}. By default, the file name is \code{"input"}. This file will exist in
this location if the post was created using \code{\link[=aoc_new_day]{aoc_new_day()}}. The initial
reading is done by \code{\link[=readLines]{readLines()}}.
}
\examples{
\dontrun{aoc_input_matrix(3, 2020)}
}
\seealso{
\code{\link[=aoc_input_vector]{aoc_input_vector()}} \code{\link[=aoc_input_data_frame]{aoc_input_data_frame()}}
}
