% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{aoc_input_vector}
\alias{aoc_input_vector}
\title{Read input as a vector}
\usage{
aoc_input_vector(day, year = NULL, mode = c("character", "numeric"))
}
\arguments{
\item{day}{An integer between 1 and 25}

\item{year}{An integer representing the year. Defaults to the current year.}

\item{mode}{Character string. One of 'character' or 'numeric'. Default is 'character'.}
}
\value{
A vector containing the puzzle input for the day and year
}
\description{
Read input as a vector
}
\examples{
\dontrun{aoc_input_vector(1, 2020, "numeric")}
}
