% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{aoc_new_conclusion}
\alias{aoc_new_conclusion}
\title{Create a conclusion post}
\usage{
aoc_new_conclusion(year = NULL)
}
\arguments{
\item{year}{An integer representing the year, from 2015 to the current year. Defaults to the current year.}
}
\value{
The path to the conclusion post (invisibly)
}
\description{
Create a conclusion post with relative path \code{./YYYY/day/conclusion} (where \code{YYYY}
is the value of \code{year}) by copying the template at \verb{./_templates/YYYY-conclusion}
and substituting \code{YYYY} for the value of \code{year}, both in the new file name and in the
file itself.
}
\examples{
\dontrun{aoc_new_introduction(2022)}
}
\seealso{
\code{\link[=aoc_delete_conclusion]{aoc_delete_conclusion()}}
}
