% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{aoc_new_intro}
\alias{aoc_new_intro}
\title{Create an introduction post}
\usage{
aoc_new_intro(year = NULL)
}
\arguments{
\item{year}{An integer representing the year, from 2015 to the current year. Defaults to the current year.}
}
\value{
The path to the introduction post (invisibly)
}
\description{
Create an introduction post with relative path \code{./YYYY/day/introduction} (where \code{YYYY}
is the value of \code{year}) by copying the template at \verb{./_templates/YYYY-intro}
and substituting \code{YYYY} for the value of \code{year}, both in the new file name and in the
file itself.
}
\examples{
\dontrun{aoc_new_intro(2022)}
}
\seealso{
\code{\link[=aoc_delete_intro]{aoc_delete_intro()}} \code{\link[=aoc_new_year]{aoc_new_year()}}
}
