% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{aoc_new_introduction}
\alias{aoc_new_introduction}
\title{Create an introduction post}
\usage{
aoc_new_introduction(year = NULL)
}
\arguments{
\item{year}{An integer representing the year. Defaults to the current year.}
}
\value{
The path to the introduction post (invisibly)
}
\description{
Create an introduction post
}
\examples{
\dontrun{aoc_new_introduction(2022)}
}
