% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{aoc_new_post}
\alias{aoc_new_post}
\alias{aoc_new_day}
\title{Set up up a new post}
\usage{
aoc_new_post(day, year = NULL)

aoc_new_day(day, year = NULL)
}
\arguments{
\item{day}{An integer between 1 and 25}

\item{year}{An integer representing the year. Defaults to the current year.}
}
\value{
Creates a new day
}
\description{
Set up up a new post

Set up up a new post
}
\examples{
\dontrun{aoc_new_post(1, 2022)}
\dontrun{aoc_new_day(1, 2022)}
}
