% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{aoc_url}
\alias{aoc_url}
\alias{aoc_input_url}
\title{Get URLs from the Advent of Code website}
\usage{
aoc_url(day, year = NULL, browse = FALSE)

aoc_input_url(day, year = NULL, browse = FALSE)
}
\arguments{
\item{day}{An integer between 1 and 25}

\item{year}{An integer representing the year. Defaults to the current year.}

\item{browse}{logical - should the URL be opened in the browser?}
}
\value{
A character string with the URL
}
\description{
Get the URL from the Advent of Code website for the puzzle and the puzzle
input for a given day and year.
}
\examples{
aoc_url(1, 2022)
aoc_input_url(1, 2022)
}
