% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-helpers.R
\name{extract_numbers}
\alias{extract_numbers}
\title{Extract all numbers from a string}
\usage{
extract_numbers(x)
}
\arguments{
\item{x}{character. The string to extract numbers from.}
}
\value{
A numeric vector containing all numbers in the string.
}
\description{
Extract all numbers from a string
}
\examples{
extract_numbers("abc123def456")
extract_numbers("Game 1:")
extract_numbers("Cards: 1 3 16 136")

}
