% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths-helpers.R
\name{GCD}
\alias{GCD}
\alias{LCM}
\title{Greatest Common Divisor (GCD) and Least Common Multiple (LCM)}
\usage{
GCD(x, y)

LCM(x, y)
}
\arguments{
\item{x}{A single integer}

\item{y}{A single integer}
}
\value{
The greatest common divisor of x and y
}
\description{
Greatest Common Divisor (GCD) and Least Common Multiple (LCM)
}
\examples{
GCD(12, 18)
GCD(12, 0)
GCD(13, 2)
LCM(12, 18)
LCM(2, 6)
LCM(3, 5)
}
