% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle-helpers.R
\name{split_at_blanks}
\alias{split_at_blanks}
\title{Split input into groups}
\usage{
split_at_blanks(lines, mode = c("character", "numeric"), split = "")
}
\arguments{
\item{lines}{A character vector}

\item{mode}{Character string. One of 'character' or 'numeric'. Default is 'character'.}

\item{split}{character. The string to split the input on. Default is \code{""}, i.e. split at blank lines}
}
\value{
A list, with one element per group originally separated by blank lines.
}
\description{
When the strings represent numbers, it's better to read them in as a character vector using \code{\link[=aoc_input_vector]{aoc_input_vector()}}, then change mode to numeric as part of call to \code{split_at_blanks()}.
}
\examples{
x <- c("123", "1234", "", "56", "567", "", "89")
split_at_blanks(x)
split_at_blanks(x, "numeric")

}
