% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{aoc_input_matrix}
\alias{aoc_input_matrix}
\title{Read puzzle input as a matrix}
\usage{
aoc_input_matrix(
  day,
  year = NULL,
  mode = c("character", "numeric"),
  split = "",
  view = FALSE
)
}
\arguments{
\item{day}{An integer between 1 and 25}

\item{year}{An integer representing the year, from 2015 to the current year. Defaults to the current year.}

\item{mode}{Character string. One of 'character' or 'numeric'. Default is 'character'.}

\item{split}{character. The string to split the input on. Default is \code{""}, i.e. one character per column}

\item{view}{logical. If \code{TRUE}, calls \code{\link[utils:View]{utils::View()}} to view the returned input.
Default is \code{FALSE}}
}
\value{
A matrix containing the puzzle input for the day and year.
}
\description{
Read in the puzzle input as a matrix, one row per line, by default one character per column.
It assumes that the input is stored in a file called \code{input} in the directory
\code{"./YYYY/day/DD"}, where \code{YYYY} and \code{DD} are the values of \code{year} and \code{day}.
This will be the case if the post was created using \code{\link[=aoc_new_day]{aoc_new_day()}}.
}
\examples{
\dontrun{aoc_input_matrix(3, 2020)}
}
\seealso{
\code{\link[=aoc_input_vector]{aoc_input_vector()}} \code{\link[=aoc_input_data_frame]{aoc_input_data_frame()}}
}
