% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{aoc_new_year}
\alias{aoc_new_year}
\title{New year}
\usage{
aoc_new_year(year = NULL, intro = TRUE)
}
\arguments{
\item{year}{An integer representing the year. Defaults to the current year.}

\item{intro}{Logical. Whether to include an introduction post. See Details.}
}
\value{
Creates a new year
}
\description{
New year
}
\examples{
\dontrun{aoc_new_year(2022)}
}
