% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths-helpers.R
\name{gcd}
\alias{gcd}
\alias{lcm}
\title{Greatest Common Divisor (GCD) and Least Common Multiple (LCM)}
\usage{
gcd(x, y)

lcm(x, y)
}
\arguments{
\item{x}{A single integer}

\item{y}{A single integer}
}
\value{
The greatest common divisor of x and y
}
\description{
Greatest Common Divisor (GCD) and Least Common Multiple (LCM)
}
\examples{
gcd(12, 18)
gcd(12, 0)
gcd(13, 2)
lcm(12, 18)
lcm(2, 6)
lcm(3, 5)
}
